+++ client/src/pages/TacticBoard/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TacticBoard";
