+++ client/src/pages/TacticBoard/translations/TacticBoard_de.json
