+++ client/src/pages/TacticBoard/translations/TacticBoard_en.json
