+++ client/src/pages/TacticBoard/translations/index.ts
@@ -1,8 +1,8 @@
 import i18next from "i18next";
-import de from "./TacticsBoard_de.json";
-import en from "./TacticsBoard_en.json";
+import de from "./TacticBoard_de.json";
+import en from "./TacticBoard_en.json";
 
-export const namespace = "TacticsBoard";
+export const namespace = "TacticBoard";
 
 i18next.addResourceBundle("de", namespace, de, true, true);
 i18next.addResourceBundle("en", namespace, en, true, true);
