--- client/src/pages/TacticsBoard/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./TacticsBoard";
